+++ packages/api-contract/src/base/Code.ts
@@ -10,15 +10,14 @@ import type { MapConstructorExec } from './types';
 
 import { SubmittableResult } from '@polkadot/api';
 import { ApiBase } from '@polkadot/api/base';
-import { assert, compactAddLength, isFunction, isUndefined, isWasm, u8aToU8a } from '@polkadot/util';
-import { blake2AsU8a } from '@polkadot/util-crypto';
+import { assert, BN_ZERO, compactAddLength, isUndefined, isWasm, u8aToU8a } from '@polkadot/util';
 
 import { Abi } from '../Abi';
 import { applyOnEvent } from '../util';
 import { Base } from './Base';
 import { Blueprint } from './Blueprint';
 import { Contract } from './Contract';
-import { createBluePrintTx, EMPTY_SALT, encodeSalt } from './util';
+import { createBluePrintTx, encodeSalt } from './util';
 
 export class CodeSubmittableResult<ApiType extends ApiTypes> extends SubmittableResult {
   public readonly blueprint?: Blueprint<ApiType>;
@@ -57,45 +56,25 @@ export class Code<ApiType extends ApiTypes> extends Base<ApiType> {
     return this.#tx;
   }
 
-  #instantiate = (constructorOrId: AbiConstructor | string | number, options: BlueprintOptions, params: CodecArg[]): SubmittableExtrinsic<ApiType, CodeSubmittableResult<ApiType>> => {
-    return isFunction(this.api.tx.contracts.instantiateWithCode)
-      ? this.#instantiateCurrent(constructorOrId, options, params)
-      : this.#instantiatePrev(constructorOrId, options, params);
-  }
-
-  #instantiateCurrent = (constructorOrId: AbiConstructor | string | number, { gasLimit = 0, salt, value = 0 }: BlueprintOptions, params: CodecArg[]): SubmittableExtrinsic<ApiType, CodeSubmittableResult<ApiType>> => {
+  #instantiate = (constructorOrId: AbiConstructor | string | number, { gasLimit = BN_ZERO, salt, value = BN_ZERO }: BlueprintOptions, params: CodecArg[]): SubmittableExtrinsic<ApiType, CodeSubmittableResult<ApiType>> => {
     return this.api.tx.contracts
-      .instantiateWithCode(value, gasLimit, compactAddLength(this.code), this.abi.findConstructor(constructorOrId).toU8a(params), encodeSalt(salt))
-      .withResultTransform(this.#transformEvents);
-  }
-
-  #instantiatePrev = (constructorOrId: AbiConstructor | string | number, { gasLimit = 0, salt, value = 0 }: BlueprintOptions, params: CodecArg[]): SubmittableExtrinsic<ApiType, CodeSubmittableResult<ApiType>> => {
-    assert(isFunction(this.api.tx.utility?.batch), 'Your chain does not include the utility pallet, for contracts v2 deployment, this is required');
-
-    const encodedSalt = encodeSalt(salt);
-    const withSalt = this.api.tx.contracts.instantiate.meta.args.length === 5;
-    const encoded = this.abi.findConstructor(constructorOrId).toU8a(params, withSalt ? EMPTY_SALT : encodedSalt);
-    const codeHash = blake2AsU8a(this.code);
-
-    return this.api.tx.utility.batch([
-      this.api.tx.contracts.putCode(compactAddLength(this.code)),
-      withSalt
-        ? this.api.tx.contracts.instantiate(value, gasLimit, codeHash, encoded, encodedSalt)
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore old style with salt included in the params itself
-        : this.api.tx.contracts.instantiate(value, gasLimit, codeHash, encoded)
-    ]).withResultTransform(this.#transformEvents);
-  }
-
-  #transformEvents = (result: ISubmittableResult): CodeSubmittableResult<ApiType> => {
-    return new CodeSubmittableResult(result, ...(applyOnEvent(result, ['CodeStored', 'Instantiated'], (records: EventRecord[]) =>
-      records.reduce(([blueprint, contract], { event }): [Blueprint<ApiType>?, Contract<ApiType>?] =>
-        this.api.events.contracts.Instantiated.is(event)
-          ? [blueprint, new Contract<ApiType>(this.api, this.abi, event.data[1], this._decorateMethod)]
-          : this.api.events.contracts.CodeStored.is(event)
-            ? [new Blueprint<ApiType>(this.api, this.abi, event.data[0], this._decorateMethod), contract]
-            : [blueprint, contract],
-      [] as [Blueprint<ApiType>?, Contract<ApiType>?])
-    ) || []));
+      .instantiateWithCode(
+        value,
+        gasLimit,
+        compactAddLength(this.code),
+        this.abi.findConstructor(constructorOrId).toU8a(params),
+        encodeSalt(salt)
+      )
+      .withResultTransform((result: ISubmittableResult) =>
+        new CodeSubmittableResult(result, ...(applyOnEvent(result, ['CodeStored', 'Instantiated'], (records: EventRecord[]) =>
+          records.reduce<[Blueprint<ApiType>?, Contract<ApiType>?]>(([blueprint, contract], { event }) =>
+            this.api.events.contracts.Instantiated.is(event)
+              ? [blueprint, new Contract<ApiType>(this.api, this.abi, event.data[1], this._decorateMethod)]
+              : this.api.events.contracts.CodeStored.is(event)
+                ? [new Blueprint<ApiType>(this.api, this.abi, event.data[0], this._decorateMethod), contract]
+                : [blueprint, contract],
+          [])
+        ) || []))
+      );
   }
 }