+++ packages/types-known/src/spec/node.ts
@@ -5,22 +5,11 @@
 
 import type { OverrideVersionedType } from '@polkadot/types/types';
 
-// these are override types for Polkadot
 const versioned: OverrideVersionedType[] = [
   {
-    minmax: [0, 260],
+    minmax: [0, undefined],
     types: {
-      AccountInfo: 'AccountInfoWithRefCount',
-      Address: 'LookupSource',
-      LookupSource: 'IndicesLookupSource'
-    }
-  },
-  {
-    minmax: [261, undefined],
-    types: {
-      AccountInfo: 'AccountInfoWithDualRefCount',
-      Address: 'MultiAddress',
-      LookupSource: 'MultiAddress'
+      // nothing, API tracks master
     }
   }
 ];