+++ packages/types/src/interfaces/definitions.ts
@@ -10,6 +10,7 @@ export { default as authorship } from './authorship/definitions';
 export { default as aura } from './aura/definitions';
 export { default as babe } from './babe/definitions';
 export { default as balances } from './balances/definitions';
+export { default as beefy } from './beefy/definitions';
 export { default as collective } from './collective/definitions';
 export { default as consensus } from './consensus/definitions';
 export { default as contracts } from './contracts/definitions';
@@ -42,9 +43,9 @@ export { default as vesting } from './vesting/definitions';
 
 // polkadot-specific types
 export { default as attestations } from './attestations/definitions';
-export { default as beefy } from './beefy/definitions';
 export { default as claims } from './claims/definitions';
 export { default as crowdloan } from './crowdloan/definitions';
+export { default as cumulus } from './cumulus/definitions';
 export { default as parachains } from './parachains/definitions';
 export { default as poll } from './poll/definitions';
 export { default as purchase } from './purchase/definitions';